week_13/day_1/polymorphism_lab/polymorphism_lab_solution_extension/src/main/java/Wallet.java
import java.util.ArrayList;

public class Wallet {
    private String name;
    private ArrayList<IScan> items;

    public Wallet(String name) {
        this.name = name;
        this.items = new ArrayList<IScan>();
    }

    public String getName() {
        return name;
    }

    public void addItem(IScan item) {
        this.items.add(item);
    }

    public int getNumberOfItems() {
        return this.items.size();
    }

    public String scanItem(int index) {
        IScan item = items.get(index);
        return item.scan();
    }
}
